#2367 fixed a bug in the csv parser but was a performance regression. This PR fixes and improves upon that regression.

Every time we were done parsing a line, because we 1. depleted a line, or 2. depleted all projected columns we search for the new line char that is not quoted. This search happened from a look behind location that was too conservative, meaning we scanned through more bytes. This PR makes sure we search from the current offset AND make sure we can skip an expensive function call if we are already on a `\n` (which is most cases).

# Performances
Most interesting, time, instructions and branch misses

## Original performance

```
         22.175,38 msec task-clock                #    9,403 CPUs utilized          
             3.010      context-switches          #    0,136 K/sec                  
                50      cpu-migrations            #    0,002 K/sec                  
         1.411.270      page-faults               #    0,064 M/sec                  
    78.397.398.336      cycles                    #    3,535 GHz                    
   108.252.556.696      instructions              #    1,38  insn per cycle         
    18.988.454.853      branches                  #  856,285 M/sec                  
       141.687.199      branch-misses             #    0,75% of all branches        

       2,358217657 seconds time elapsed

      19,419798000 seconds user
       2,752847000 seconds sys
```

## Regression

```
         25.706,17 msec task-clock                #    9,680 CPUs utilized          
             3.893      context-switches          #    0,151 K/sec                  
                42      cpu-migrations            #    0,002 K/sec                  
         1.411.335      page-faults               #    0,055 M/sec                  
    91.945.785.516      cycles                    #    3,577 GHz                    
   147.508.250.208      instructions              #    1,60  insn per cycle         
    26.256.706.556      branches                  # 1021,417 M/sec                  
       158.175.002      branch-misses             #    0,60% of all branches        

       2,655604035 seconds time elapsed

      22,992119000 seconds user
       2,711204000 seconds sys
```

## Improvement

```
         21.093,12 msec task-clock                #    9,459 CPUs utilized          
             2.457      context-switches          #    0,116 K/sec                  
                88      cpu-migrations            #    0,004 K/sec                  
         1.411.266      page-faults               #    0,067 M/sec                  
    74.342.626.979      cycles                    #    3,524 GHz                    
   104.044.686.745      instructions              #    1,40  insn per cycle         
    17.945.290.110      branches                  #  850,765 M/sec                  
       164.631.548      branch-misses             #    0,92% of all branches        

       2,230055835 seconds time elapsed

      18,539514000 seconds user
       2,553771000 seconds sys
```